Move JavacProblem & JavacProblemConverter to internal package
